Updated Vertex code to match the package renaming
